member 타입 long으로 변경 및 메세지 db저장 구현 중

+++ insider/src/main/java/com/kh/insider/vo/MemberMessageVO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Data @Builder @NoArgsConstructor @AllArgsConstructor
 public class MemberMessageVO {
 	
-	private int memberNo;
+	private long memberNo;
 	private String memberName, content;
 	private long time;
